# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/shops/{shopId}/work-orders": {
      "post": {
        "operationId": "WorkOrdersController_create",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OT creada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Abre una orden de trabajo (queda en draft con número OT-XXXX).",
        "tags": [
          "work-orders"
        ]
      },
      "get": {
        "operationId": "WorkOrdersController_list",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "awaiting_approval",
                "approved",
                "in_progress",
                "completed",
                "invoiced",
                "paid",
                "cancelled"
              ]
            }
          },
          {
            "name": "customer_id",
            "required": false,
            "in": "query",
            "description": "Historial de un cliente.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicle_id",
            "required": false,
            "in": "query",
            "description": "Historial de un vehículo.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Número de OT: `12` u `OT-0012`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de OT, de la más nueva a la más vieja.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderViewPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista OT del taller (filtros: estado, cliente, vehículo; búsqueda por número).",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}": {
      "get": {
        "operationId": "WorkOrdersController_get",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetailView"
                }
              }
            }
          },
          "404": {
            "description": "No existe en este taller."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalle de una OT con sus líneas y totales.",
        "tags": [
          "work-orders"
        ]
      },
      "patch": {
        "operationId": "WorkOrdersController_update",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderView"
                }
              }
            }
          },
          "409": {
            "description": "La OT ya está cerrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita la cabecera de una OT abierta.",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/items": {
      "post": {
        "operationId": "WorkOrdersController_addItem",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OT con sus líneas y totales actualizados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetailView"
                }
              }
            }
          },
          "409": {
            "description": "La OT ya está cerrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agrega una línea (mano de obra o pieza); recalcula totales con ITBIS.",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/items/{itemId}": {
      "patch": {
        "operationId": "WorkOrdersController_updateItem",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetailView"
                }
              }
            }
          },
          "409": {
            "description": "La OT ya está cerrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita una línea; recalcula totales.",
        "tags": [
          "work-orders"
        ]
      },
      "delete": {
        "operationId": "WorkOrdersController_removeItem",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OT con sus líneas y totales actualizados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetailView"
                }
              }
            }
          },
          "409": {
            "description": "La OT ya está cerrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Quita una línea; recalcula totales.",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/v1/shops/{shopId}/work-orders/{workOrderId}/transitions": {
      "post": {
        "operationId": "WorkOrdersController_transition",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OT con el nuevo estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDetailView"
                }
              }
            }
          },
          "409": {
            "description": "Transición no permitida desde el estado actual, u OT sin líneas."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambia el estado: draft → in_progress → completed; cancelled desde cualquier estado abierto. Completar exige al menos una línea. Al completar se avisa al cliente.",
        "tags": [
          "work-orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "MechMate API",
    "description": "Backend del ecosistema MechMate (ERP de talleres + marketplace de repuestos, RD). Dinero en centavos como string (`Cents`); fechas ISO-8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWorkOrderDto": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string"
          },
          "vehicle_id": {
            "type": "string",
            "description": "Debe ser un vehículo de ese cliente."
          },
          "complaint": {
            "type": "string",
            "example": "Ruido al frenar y vibración en el volante",
            "minLength": 3,
            "maxLength": 2000
          },
          "notes": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          },
          "mileage_in": {
            "type": "object",
            "example": 185000,
            "nullable": true,
            "description": "Odómetro al recibir el vehículo."
          },
          "assigned_member_id": {
            "type": "object",
            "nullable": true,
            "description": "shop_members.id del mecánico asignado."
          },
          "promised_at": {
            "type": "object",
            "example": "2026-10-02T17:00:00-04:00",
            "nullable": true
          }
        },
        "required": [
          "customer_id",
          "vehicle_id",
          "complaint"
        ]
      },
      "UpdateWorkOrderDto": {
        "type": "object",
        "properties": {
          "complaint": {
            "type": "string",
            "example": "Ruido al frenar y vibración en el volante",
            "minLength": 3,
            "maxLength": 2000
          },
          "notes": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          },
          "mileage_in": {
            "type": "object",
            "example": 185000,
            "nullable": true,
            "description": "Odómetro al recibir el vehículo."
          },
          "assigned_member_id": {
            "type": "object",
            "nullable": true,
            "description": "shop_members.id del mecánico asignado."
          },
          "promised_at": {
            "type": "object",
            "example": "2026-10-02T17:00:00-04:00",
            "nullable": true
          }
        }
      },
      "AddItemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "labor",
              "part"
            ],
            "example": "labor",
            "description": "labor = mano de obra; part = pieza."
          },
          "description": {
            "type": "string",
            "example": "Cambio de pastillas delanteras",
            "maxLength": 300
          },
          "part_number": {
            "type": "object",
            "example": "04465-02220",
            "nullable": true,
            "maxLength": 60
          },
          "quantity": {
            "type": "string",
            "example": "1.5",
            "description": "Hasta 3 decimales (horas, unidades, galones…)."
          },
          "unit_price_cents": {
            "type": "string",
            "example": "120000",
            "description": "Centavos como string: \"120000\" = RD$1,200.00."
          },
          "tax_rate_bps": {
            "type": "number",
            "example": 1800,
            "default": 1800,
            "description": "ITBIS en basis points (1800 = 18%, 0 = exento)."
          },
          "requires_approval": {
            "type": "boolean",
            "default": false,
            "description": "true = línea propuesta: el cliente la aprueba o rechaza desde el enlace de la DVI."
          }
        },
        "required": [
          "type",
          "description",
          "quantity",
          "unit_price_cents"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "labor",
              "part"
            ],
            "example": "labor",
            "description": "labor = mano de obra; part = pieza."
          },
          "description": {
            "type": "string",
            "example": "Cambio de pastillas delanteras",
            "maxLength": 300
          },
          "part_number": {
            "type": "object",
            "example": "04465-02220",
            "nullable": true,
            "maxLength": 60
          },
          "quantity": {
            "type": "string",
            "example": "1.5",
            "description": "Hasta 3 decimales (horas, unidades, galones…)."
          },
          "unit_price_cents": {
            "type": "string",
            "example": "120000",
            "description": "Centavos como string: \"120000\" = RD$1,200.00."
          },
          "tax_rate_bps": {
            "type": "number",
            "example": 1800,
            "default": 1800,
            "description": "ITBIS en basis points (1800 = 18%, 0 = exento)."
          },
          "requires_approval": {
            "type": "boolean",
            "default": false,
            "description": "true = línea propuesta: el cliente la aprueba o rechaza desde el enlace de la DVI."
          }
        }
      },
      "TransitionDto": {
        "type": "object",
        "properties": {
          "to": {
            "type": "string",
            "enum": [
              "in_progress",
              "completed",
              "cancelled"
            ],
            "example": "in_progress"
          },
          "reason": {
            "type": "object",
            "example": "El cliente decidió no reparar",
            "nullable": true,
            "maxLength": 500,
            "description": "Solo para cancelled."
          }
        },
        "required": [
          "to"
        ]
      },
...
      "WorkOrderView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "number": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkOrderStatus"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummary"
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleSummary"
          },
          "complaint": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "mileage_in": {
            "type": "number",
            "nullable": true
          },
          "assigned_member_id": {
            "type": "string",
            "nullable": true
          },
          "promised_at": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "subtotal_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "tax_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "total_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "started_at": {
            "type": "string",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "nullable": true
          },
          "cancelled_at": {
            "type": "string",
            "nullable": true
          },
          "cancellation_reason": {
            "type": "string",
            "nullable": true
          },
          "created_by_account_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "number",
          "code",
          "status",
          "customer",
          "vehicle",
          "complaint",
          "notes",
          "mileage_in",
          "assigned_member_id",
          "promised_at",
          "currency",
          "subtotal_cents",
          "tax_cents",
          "total_cents",
          "started_at",
          "completed_at",
          "cancelled_at",
          "cancellation_reason",
          "created_by_account_id",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "WorkOrderStatus": {
        "type": "string",
        "enum": [
          "draft",
          "awaiting_approval",
          "approved",
          "in_progress",
          "completed",
          "invoiced",
          "paid",
          "cancelled"
        ]
      },
      "CustomerSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "full_name",
          "phone"
        ],
        "additionalProperties": false
      },
      "VehicleSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "number",
            "nullable": true
          },
          "plate": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "make",
          "model",
          "year",
          "plate"
        ],
        "additionalProperties": false
      },
      "Cents": {
        "type": "string"
      },
      "WorkOrderDetailView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "number": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkOrderStatus"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummary"
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleSummary"
          },
          "complaint": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "mileage_in": {
            "type": "number",
            "nullable": true
          },
          "assigned_member_id": {
            "type": "string",
            "nullable": true
          },
          "promised_at": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "subtotal_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "tax_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "total_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "started_at": {
            "type": "string",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "nullable": true
          },
          "cancelled_at": {
            "type": "string",
            "nullable": true
          },
          "cancellation_reason": {
            "type": "string",
            "nullable": true
          },
          "created_by_account_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkOrderItemView"
            }
          }
        },
        "required": [
          "assigned_member_id",
          "cancellation_reason",
          "cancelled_at",
          "code",
          "complaint",
          "completed_at",
          "created_at",
          "created_by_account_id",
          "currency",
          "customer",
          "id",
          "items",
          "mileage_in",
          "notes",
          "number",
          "promised_at",
          "started_at",
          "status",
          "subtotal_cents",
          "tax_cents",
          "total_cents",
          "updated_at",
          "vehicle"
        ],
        "additionalProperties": false
      },
...
      "WorkOrderViewPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkOrderView"
            }
          },
          "next_cursor": {
            "type": "string",
            "nullable": true,
            "description": "Pasar como `cursor` para la siguiente página; null si no hay más."
          }
        },
        "required": [
          "items",
          "next_cursor"
        ]
      },
...
    }
  }
}